src/lymphocyte/pyproject.toml
[tool.isort]
profile = "black"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.poetry]
name = "lymphocyte"
version = "0.1.0"
description = ""
authors = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
dependency-injector = {version = "4.41.0", extras = ["yaml"]}
fastapi = "0.103.1"
httpx = "0.25.0"
hypercorn = "0.14.4"
jinja2 = "3.1.2"
prometheus-client = "0.17.1"
prometheus-fastapi-instrumentator = "6.1.0"
pydantic = "2.1.1"
pydantic-settings = "2.0.2"
pydantic-core = "2.4.0"
pytoml = "0.1.21"
pyyaml = "6.0.1"
rich = "13.5.3"
websockets = "^12.0"
psutil = "5.9.8"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
freezegun = "^1.2.2"
isort = "^5.12.0"
mypy = "^1.4.1"
mypy-extensions = "^1.0.0"
pylint = "^3.2"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
types-colorama = "^0.4.15.20240311"
types-jinja2 = "^2.11.9"
types-pygments = "^2.18.0.20240506"
types-pyyaml = "^6.0.12.20240311"
types-requests = "^2.32.0.20240712"
types-setuptools = "^70.3.0.20240710"
types-ujson = "^5.10.0.20240515"
types-psutil = "^6.0.0.20240621"
watchdog = "^3.0.0"
lxml = "^5.2.2"
pdoc = "^14.7.0"
bandit = {extras = ["toml"], version = "^1.8.3"}
pylic = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pylint.main]
load-plugins = [
    "pylint.extensions.code_style",
    "pylint.extensions.check_elif",
    "pylint.extensions.confusing_elif",
    "pylint.extensions.mccabe",
    "pylint.extensions.empty_comment",
    "pylint.extensions.private_import",
    "pylint.extensions.magic_value",
    "pylint.extensions.redefined_variable_type",
    "pylint.extensions.typing",
]
fail-on = "useless-suppression"

[tool.pylint.messages_control]
disable = "all"
enable = [
    "F", # Fatal
    "E", # Error
    "W", # Warning
    # "C", # Convention
    # "R", # Refactor
    # "I", # Information

    "unused-import",

    # From pylint.extensions.typing
    "broken-noreturn",
    "broken-collections-callable",
    "deprecated-typing-alias",
    "consider-using-alias",
    "consider-alternative-union-syntax",
    "redundant-typehint-argument",
]

[tool.mypy]
warn_unreachable = true

# Strict mode enables the following flags:
warn_unused_configs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_reexport = true
strict_equality = true
extra_checks = true

[tool.bandit]
exclude_dirs = [
    ".venv",
    "tests",
]

[tool.pylic]
safe_licenses = [
    "Apache Software License",
    "MIT License",
    "Python Software Foundation License",
    "Mozilla Public License 2.0 (MPL 2.0)",
    "ISC",
    "BSD License",
    "GNU Lesser General Public License v2 (LGPLv2)",
    "GNU General Public License v2 (GPLv2)",
    "Public Domain",
]
unsafe_packages = [
]
ignore_packages = [
]
